Added text renderer

+++ src/graphics/Glyph.h
@@ -2,9 +2,9 @@
 #define VOX_GLYPH_H
 
 struct Glyph {
-    glm::ivec2 advance;
-    glm::ivec2 size;
-    glm::ivec2 bearing;
+    glm::vec2 advance;
+    glm::vec2 size;
+    glm::vec2 bearing;
     glm::vec2 uv;
 };
